requirements.txt
attrs==24.2.0
aws-cdk-lib==2.151.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
aws-cdk.assets==1.204.0
aws-cdk.aws-acmpca==1.204.0
aws-cdk.aws-apigateway==1.204.0
aws-cdk.aws-applicationautoscaling==1.204.0
aws-cdk.aws-autoscaling==1.204.0
aws-cdk.aws-autoscaling-common==1.204.0
aws-cdk.aws-autoscaling-hooktargets==1.204.0
aws-cdk.aws-certificatemanager==1.204.0
aws-cdk.aws-cloudformation==1.204.0
aws-cdk.aws-cloudfront==1.204.0
aws-cdk.aws-cloudwatch==1.204.0
aws-cdk.aws-codebuild==1.204.0
aws-cdk.aws-codecommit==1.204.0
aws-cdk.aws-codeguruprofiler==1.204.0
aws-cdk.aws-codepipeline==1.204.0
aws-cdk.aws-codestarnotifications==1.204.0
aws-cdk.aws-cognito==1.204.0
aws-cdk.aws-ec2==1.204.0
aws-cdk.aws-ecr==1.204.0
aws-cdk.aws-ecr-assets==1.204.0
aws-cdk.aws-ecs==1.204.0
aws-cdk.aws-ecs-patterns==1.204.0
aws-cdk.aws-efs==1.204.0
aws-cdk.aws-elasticloadbalancing==1.204.0
aws-cdk.aws-elasticloadbalancingv2==1.204.0
aws-cdk.aws-events==1.204.0
aws-cdk.aws-events-targets==1.204.0
aws-cdk.aws-globalaccelerator==1.204.0
aws-cdk.aws-iam==1.204.0
aws-cdk.aws-kinesis==1.204.0
aws-cdk.aws-kinesisfirehose==1.204.0
aws-cdk.aws-kms==1.204.0
aws-cdk.aws-lambda==1.204.0
aws-cdk.aws-logs==1.204.0
aws-cdk.aws-route53==1.204.0
aws-cdk.aws-route53-targets==1.204.0
aws-cdk.aws-s3==1.204.0
aws-cdk.aws-s3-assets==1.204.0
aws-cdk.aws-sam==1.204.0
aws-cdk.aws-secretsmanager==1.204.0
aws-cdk.aws-servicediscovery==1.204.0
aws-cdk.aws-signer==1.204.0
aws-cdk.aws-sns==1.204.0
aws-cdk.aws-sns-subscriptions==1.204.0
aws-cdk.aws-sqs==1.204.0
aws-cdk.aws-ssm==1.204.0
aws-cdk.aws-stepfunctions==1.204.0
aws-cdk.cloud-assembly-schema==1.204.0
aws-cdk.core==1.204.0
aws-cdk.custom-resources==1.204.0
aws-cdk.cx-api==1.204.0
aws-cdk.region-info==1.204.0
boto3==1.35.10
botocore==1.35.10
cattrs==23.2.3
constructs==10.3.0
exceptiongroup==1.2.2
importlib_resources==6.4.4
jmespath==1.0.1
jsii==1.103.1
publication==0.0.3
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==1.26.20
zipp==3.20.1
